Create moveToMethod At Player Class
